requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
av==14.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ctranslate2==4.5.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.3.0
google-auth==2.38.0
google-genai==1.7.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
llvmlite==0.44.0
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
onnxruntime==1.21.0
openai-whisper==20240930
packaging==24.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.50
protobuf==6.30.1
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==309
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
srt==3.5.3
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
vosk==0.3.45
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
youtube-transcript-api==1.0.2
